massage in swissroll so it fits nicer with other tasks
