+++ src/bmi/samplers/api.py
@@ -2,22 +2,12 @@ from bmi.samplers.additive_uniform import AdditiveUniformSampler
 from bmi.samplers.dispersion import parametrised_correlation_matrix
 from bmi.samplers.split_student_t import SplitStudentT
 from bmi.samplers.splitmultinormal import BivariateNormalSampler, SplitMultinormal
-
-# isort: off
-from bmi.samplers.transformed import (
-    BivariateUniformMarginsSampler,
-    SwissRollSampler,
-    TransformedSampler,
-)
-
-# isort: on
+from bmi.samplers.transformed import TransformedSampler
 
 __all__ = [
     "AdditiveUniformSampler",
     "parametrised_correlation_matrix",
     "BivariateNormalSampler",
-    "BivariateUniformMarginsSampler",
-    "SwissRollSampler",
     "SplitMultinormal",
     "SplitStudentT",
     "TransformedSampler",
