+++ src/bmi/transforms/api.py
@@ -1,6 +1,6 @@
 from bmi.transforms.invert_cdf import invert_cdf
 from bmi.transforms.rotate import Spiral, skew_symmetrize, so_generator
-from bmi.transforms.simple import half_cube, normal_cdf
+from bmi.transforms.simple import half_cube, normal_cdf, swissroll2d
 
 __all__ = [
     "invert_cdf",
@@ -9,4 +9,5 @@ __all__ = [
     "skew_symmetrize",
     "normal_cdf",
     "half_cube",
+    "swissroll2d",
 ]
